backend/pyproject.toml
[tool.poetry]
name = "juddges"
version = "0.1.0"
description = "Juddges App - AI-powered judicial decision search for Polish and UK courts"
authors = ["Lukasz Augustyniak <aisolutions@lukaszaugustyniak.com>"]
readme = "README.md"
packages = [
    { include = "app" },
    { include = "packages/juddges_search" },
    { include = "packages/schema_generator_agent" },
]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
uvicorn = "^0.23.2"
gunicorn = "^23.0.0"
langserve = { extras = ["server"], version = "^0.3.2" }
langchain = "^0.3"
langchain-community = "^0.3.19"
langchain-openai = "^0.3.7"
juddges_search = { path = "packages/juddges_search", develop = true }
schema_generator_agent = { path = "packages/schema_generator_agent", develop = true }
langfuse = "^3.2.3"
aiohttp = "^3.11.14"
httpx = "^0.27.0"
supabase = "^2.20.0"
langgraph-checkpoint-postgres = "^2.0.23"
psycopg-binary = "^3.2.10"
celery = { extras = ["redis"], version = "^5.5.3" }
werkzeug = "^3.1.3"
slowapi = "^0.1.9"
redis = "^5.2.0"
pydantic = "<2.11"
litellm = "^1.78.0"
pandas = "^2.2.0"
openpyxl = "^3.1.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.291.3"}

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.36"
jupyterlab = "^4.3.6"
ruff = "^0.11.0"
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
poethepoet = "^0.29.0"

[tool.poe.tasks]
# Test commands
test = "pytest tests/ -v"
test-unit = "pytest tests/ -v -m unit"
test-integration = "pytest tests/ -v -m integration"
test-cov = "pytest tests/ --cov=juddges_search --cov-report=html --cov-report=term-missing"
test-cov-html = "pytest tests/ --cov=juddges_search --cov-report=html"

# Code quality commands
lint = "ruff check ."
lint-fix = "ruff check . --fix"
format = "ruff format ."
format-check = "ruff format . --check"

# Combined commands
check = ["lint", "format-check"]
check-all = ["lint", "format-check", "test"]
ci = ["lint", "format-check"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-ra -q --strict-markers"
filterwarnings = [
    "ignore:'enablePackrat' deprecated - use 'enable_packrat':DeprecationWarning:pyiceberg\\.expressions\\.parser",
    "ignore:'escChar' argument is deprecated, use 'esc_char':DeprecationWarning:pyiceberg\\.expressions\\.parser",
    "ignore:'unquoteResults' argument is deprecated, use 'unquote_results':DeprecationWarning:pyiceberg\\.expressions\\.parser",
    "ignore:Using `@model_validator` with mode='after' on a classmethod is deprecated.*:DeprecationWarning:pyiceberg\\.table\\.metadata",
]
markers = [
    "integration: marks tests as integration tests (require external services)",
    "unit: marks tests as unit tests",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "C4", "SIM", "RET", "PTH", "PERF", "B", "S", "RUF", "TC", "FA"]
ignore = [
    "E501",   # line too long (handled by formatter)
    "S101",   # assert usage (needed in tests)
    "B008",   # function call in default arg (FastAPI Depends pattern)
    "B904",   # raise from in except (TODO: fix incrementally)
    "PTH123", # open() vs Path.open() (style preference)
    "S311",   # random module (non-security uses are fine)
    "PERF401", # list comprehension (TODO: fix incrementally)
    "S110",    # try-except-pass (TODO: add logging incrementally)
    "RUF012",  # mutable class variable (Pydantic models use this pattern)
    "S608",    # SQL string concat (only in admin scripts)
    "PTH118",  # os.path.join (sys.path manipulation needs os.path)
    "PTH120",  # os.path.dirname (sys.path manipulation needs os.path)
    "S314",    # XML parsing (xml_converter uses lxml safely)
    "S104",    # bind all interfaces (intentional for Docker)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["S101", "S106", "SIM102"]
"scripts/**/*.py" = ["S608"]
"examples/**/*.py" = ["S113"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
